Separated audio player functions.
Put them into util file.
