+++ src/client/components/AudioPlayer.jsx
@@ -1,33 +1,12 @@
 import React, {Component} from 'react';
-
+import { setVolume, makeItPlay } from '../util/ClientFunctions.jsx';
 
 class AudioPlayer extends Component {
   constructor(props) {
     super(props);
 
-    this.makeItPlay = this.makeItPlay.bind(this);
-    this.setVolume = this.setVolume.bind(this);
-  }
-
-  makeItPlay() {
-    let player = document.getElementById("player")
-
-    if (this.props.stationFeed.name == '') {
-      this.props.seekStation();
-    } else if (player.paused) {
-      player.play();
-      pButton.className = "";
-      pButton.className = "fa fa-pause fa-2x";
-    } else {
-      player.pause();
-      pButton.className = "";
-      pButton.className = "fa fa-play fa-2x";
-    }
-  }
-
-  setVolume(event) {
-    let player = document.getElementById("player");
-    player.volume = event.target.value / 100
+    this.makeItPlay = makeItPlay.bind(this);
+    this.setVolume = setVolume.bind(this);
   }
 
   render() {
@@ -47,4 +26,5 @@ class AudioPlayer extends Component {
     );
   }
 }
+
 export default AudioPlayer;
